requirements.txt
streamlit==1.28.0
langchain==0.0.350
langchain-openai==0.0.5
psycopg2-binary==2.9.7
pandas==2.1.3
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23